created the repository for the Traffic Image API
