--- app/src/main/java/com/sanathcoding/sglivetrafficimage/core/constValue/ConstValue.kt
@@ -1,10 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.core.constValue
-
-object ConstValue {
-    const val MIN_NAME_LENGTH = 3
-    const val MAX_NAME_LENGTH = 20
-    const val PREFERENCES_NAME = "name_preferences"
-    const val USER_NAME = "userName"
-    const val LOGIN = "login"
-    const val CAMERA_DETAIL = "camera_detail"
-}
