+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/remote/TrafficImageApi.kt
@@ -1,0 +1,10 @@
+package com.sanathcoding.sglivetrafficimage.core.data.remote
+
+import com.sanathcoding.sglivetrafficimage.core.data.remote.dto.TrafficImageDto
+import retrofit2.http.GET
+
+interface TrafficImageApi {
+
+    @GET("/v1/transport/traffic-images")
+    suspend fun getTrafficCameras(): TrafficImageDto
+}
