+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/remote/dto/ApiInfo.kt
@@ -1,0 +1,5 @@
+package com.sanathcoding.sglivetrafficimage.core.data.remote.dto
+
+data class ApiInfo(
+    val status: String
+)
