+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/remote/dto/ImageMetadata.kt
@@ -1,0 +1,7 @@
+package com.sanathcoding.sglivetrafficimage.core.data.remote.dto
+
+data class ImageMetadata(
+    val height: Int,
+    val md5: String,
+    val width: Int
+)
