+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/remote/dto/Item.kt
@@ -1,0 +1,6 @@
+package com.sanathcoding.sglivetrafficimage.core.data.remote.dto
+
+data class Item(
+    val cameras: List<Camera>,
+    val timestamp: String
+)
