+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/remote/dto/Location.kt
@@ -1,0 +1,6 @@
+package com.sanathcoding.sglivetrafficimage.core.data.remote.dto
+
+data class Location(
+    val latitude: Double,
+    val longitude: Double
+)
