+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/remote/dto/TrafficImageDto.kt
@@ -1,0 +1,9 @@
+package com.sanathcoding.sglivetrafficimage.core.data.remote.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class TrafficImageDto(
+    @SerializedName("api_info")
+    val apiInfo: ApiInfo,
+    val items: List<Item>
+)
