+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/navigation/SetUpNavGraph.kt
@@ -4,9 +4,9 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.sanathcoding.sglivetrafficimage.presentation.camera_details_screen.CameraDetailScreen
-import com.sanathcoding.sglivetrafficimage.presentation.home_screen.NewHome
-import com.sanathcoding.sglivetrafficimage.feature_login.presentation.loginScreen.LoginScreen
+import com.sanathcoding.sglivetrafficimage.map_feature.presentation.camera_details_screen.CameraDetailScreen
+import com.sanathcoding.sglivetrafficimage.map_feature.presentation.home_screen.NewHome
+import com.sanathcoding.sglivetrafficimage.login_feature.presentation.loginScreen.LoginScreen
 
 const val TEST_ROUTE = "test_route"
