--- app/src/main/java/com/sanathcoding/sglivetrafficimage/feature_login/presentation/loginScreen/ValidateEvent.kt
@@ -1,7 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.feature_login.presentation.loginScreen
-
-sealed interface ValidateEvent {
-    object Success: ValidateEvent
-    object DataAdded: ValidateEvent
-    object PasswordNotMatch: ValidateEvent
-}
