+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/login_feature/domain/model/UserCredential.kt
@@ -1,0 +1,7 @@
+package com.sanathcoding.sglivetrafficimage.login_feature.domain.model
+
+data class UserCredential(
+    val userName: String,
+    val password: String,
+    val id: Int? = null
+)
