+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/login_feature/presentation/loginScreen/ValidateEvent.kt
@@ -1,0 +1,7 @@
+package com.sanathcoding.sglivetrafficimage.login_feature.presentation.loginScreen
+
+sealed interface ValidateEvent {
+    object Success: ValidateEvent
+    object DataAdded: ValidateEvent
+    object PasswordNotMatch: ValidateEvent
+}
