+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/map_feature/domain/repository/TrafficImageRepository.kt
@@ -1,0 +1,9 @@
+package com.sanathcoding.sglivetrafficimage.map_feature.domain.repository
+
+import com.sanathcoding.sglivetrafficimage.core.data.remote.dto.TrafficImageDto
+
+interface TrafficImageRepository {
+
+    suspend fun getCameras(): TrafficImageDto
+
+}
