+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/map_feature/presentation/home_screen/screen/MapScreen.kt
@@ -1,0 +1,22 @@
+package com.sanathcoding.sglivetrafficimage.map_feature.presentation.home_screen.screen
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun MapScreen() {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = "Map Screen",
+            fontSize = MaterialTheme.typography.h3.fontSize
+        )
+    }
+}
