--- app/src/main/java/com/sanathcoding/sglivetrafficimage/presentation/camera_details_screen/CameraDetailScreen.kt
@@ -1,25 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.presentation.camera_details_screen
-
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-@Composable
-fun CameraDetailScreen(
-    cameraId: String,
-) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(
-            text = "Number is $cameraId",
-            fontSize = MaterialTheme.typography.h3.fontSize,
-        )
-    }
-}
